add validate date in range and in the past year helpers
